+++ src/swap/central/godex.ts
@@ -31,7 +31,13 @@ import {
   makeSwapPluginQuote,
   SwapOrder
 } from '../../util/swapHelpers'
-import { convertRequest, getAddress, memoType } from '../../util/utils'
+import {
+  convertRequest,
+  denominationToNative,
+  getAddress,
+  memoType,
+  nativeToDenomination
+} from '../../util/utils'
 import { asNumberString, EdgeSwapRequestPlugin } from '../types'
 
 const pluginId = 'godex'
@@ -208,13 +214,15 @@ export function makeGodexPlugin(opts: EdgeCorePluginOptions): EdgeSwapPlugin {
     // Convert the native amount to a denomination:
     const quoteAmount =
       request.quoteFor === 'from'
-        ? await request.fromWallet.nativeToDenomination(
+        ? nativeToDenomination(
+            request.fromWallet,
             request.nativeAmount,
-            request.fromCurrencyCode
+            request.fromTokenId
           )
-        : await request.toWallet.nativeToDenomination(
+        : nativeToDenomination(
+            request.toWallet,
             request.nativeAmount,
-            request.toCurrencyCode
+            request.toTokenId
           )
 
     // Swap the currencies if we need a reverse quote:
@@ -243,13 +251,15 @@ export function makeGodexPlugin(opts: EdgeCorePluginOptions): EdgeSwapPlugin {
     // min_amount returned could be for a different network than the user is requesting.
     // Check the minimum:
     const nativeMin = reverseQuote
-      ? await request.toWallet.denominationToNative(
+      ? denominationToNative(
+          request.toWallet,
           reply.min_amount,
-          request.toCurrencyCode
+          request.toTokenId
         )
-      : await request.fromWallet.denominationToNative(
+      : denominationToNative(
+          request.fromWallet,
           reply.min_amount,
-          request.fromCurrencyCode
+          request.fromTokenId
         )
 
     if (lt(request.nativeAmount, nativeMin)) {
@@ -298,16 +308,18 @@ export function makeGodexPlugin(opts: EdgeCorePluginOptions): EdgeSwapPlugin {
     log('sendReply' + JSON.stringify(sendReply, null, 2))
     const quoteInfo = asQuoteInfo(sendReply)
     const fromNativeAmount = floor(
-      await request.fromWallet.denominationToNative(
+      denominationToNative(
+        request.fromWallet,
         quoteInfo.deposit_amount,
-        request.fromCurrencyCode
+        request.fromTokenId
       ),
       0
     )
     const toNativeAmount = floor(
-      await request.toWallet.denominationToNative(
+      denominationToNative(
+        request.toWallet,
         quoteInfo.withdrawal_amount,
-        request.toCurrencyCode
+        request.toTokenId
       ),
       0
     )