+++ src/swap/defi/thorchain/swapkit.ts
@@ -29,9 +29,11 @@ import {
 } from '../../../util/swapHelpers'
 import {
   convertRequest,
+  denominationToNative,
   fetchInfo,
   fetchWaterfall,
   getAddress,
+  nativeToDenomination,
   promiseWithTimeout
 } from '../../../util/utils'
 import { EdgeSwapRequestPlugin } from '../../types'
@@ -248,9 +250,10 @@ export function makeSwapKitPlugin(opts: EdgeCorePluginOptions): EdgeSwapPlugin {
       throw new SwapCurrencyError(swapInfo, request)
     }
 
-    const sellAmount = await fromWallet.nativeToDenomination(
+    const sellAmount = nativeToDenomination(
+      fromWallet,
       nativeAmount,
-      fromCurrencyCode
+      fromTokenId
     )
 
     const quoteParams: ThorSwapQuoteParams = {
@@ -322,7 +325,7 @@ export function makeSwapKitPlugin(opts: EdgeCorePluginOptions): EdgeSwapPlugin {
     const { expectedBuyAmount, providers, targetAddress, expiration } = thorSwap
 
     const toNativeAmount = toFixed(
-      await toWallet.denominationToNative(expectedBuyAmount, toCurrencyCode),
+      denominationToNative(toWallet, expectedBuyAmount, toTokenId),
       0,
       0
     )