+++ test/testpartner.ts
@@ -14,6 +14,7 @@ import {
   EdgeCurrencyWallet,
   EdgeSwapQuote,
   EdgeSwapRequest,
+  EdgeTokenId,
   lockEdgeCorePlugins,
   makeEdgeContext,
   makeFakeEdgeWorld
@@ -23,6 +24,7 @@ import currencyPlugins from 'edge-currency-plugins'
 import fs from 'fs'
 
 import edgeExchangePlugins from '../src'
+import { denominationToNative } from '../src/util/utils'
 import { arrrCurrencyInfo } from './fakeArrrInfo'
 import { makeFakePlugin } from './fakeCurrencyPlugin'
 import { asTestConfig } from './testconfig'
@@ -49,6 +51,8 @@ interface FetchQuoteParams {
   toWallet: EdgeCurrencyWallet
   fromCurrencyCode: string
   toCurrencyCode: string
+  fromTokenId: EdgeTokenId
+  toTokenId: EdgeTokenId
   quoteFor: 'from' | 'to'
   exchangeAmount?: string
   nativeAmount?: string
@@ -165,6 +169,8 @@ async function main(): Promise<void> {
     toWallet,
     fromCurrencyCode,
     toCurrencyCode,
+    fromTokenId,
+    toTokenId,
     quoteFor,
     exchangeAmount,
     nativeAmount
@@ -173,15 +179,13 @@ async function main(): Promise<void> {
     if (exchangeAmount != null) {
       console.log(`Amount: ${quoteFor} ${exchangeAmount}`)
       if (quoteFor === 'from') {
-        nativeAmount = await fromWallet.denominationToNative(
+        nativeAmount = denominationToNative(
+          fromWallet,
           exchangeAmount,
-          fromCurrencyCode
+          fromTokenId
         )
       } else {
-        nativeAmount = await toWallet.denominationToNative(
-          exchangeAmount,
-          toCurrencyCode
-        )
+        nativeAmount = denominationToNative(toWallet, exchangeAmount, toTokenId)
       }
     } else if (nativeAmount != null) {
       console.log(`Request: ${fromCurrencyCode} to ${toCurrencyCode}`)
@@ -189,22 +193,6 @@ async function main(): Promise<void> {
     } else {
       throw new Error('No nativeAmount or exchangeAmount')
     }
-
-    const getTokenId = (
-      coreWallet: EdgeCurrencyWallet,
-      currencyCode: string
-    ): string | null => {
-      if (coreWallet.currencyInfo.currencyCode === currencyCode) return null
-      const { allTokens } = coreWallet.currencyConfig
-      return (
-        Object.keys(allTokens).find(
-          edgeToken => allTokens[edgeToken].currencyCode === currencyCode
-        ) ?? null
-      )
-    }
-
-    const fromTokenId = getTokenId(fromWallet, fromCurrencyCode)
-    const toTokenId = getTokenId(toWallet, toCurrencyCode)
 
     const request: EdgeSwapRequest = {
       fromWallet,
@@ -230,8 +218,10 @@ async function main(): Promise<void> {
   await fetchQuote({
     fromWallet: btcWallet,
     fromCurrencyCode: 'BTC',
+    fromTokenId: null,
     toWallet: ethWallet,
     toCurrencyCode: 'ETH',
+    toTokenId: null,
     exchangeAmount: '0.002',
     quoteFor: 'from'
   })
@@ -239,8 +229,10 @@ async function main(): Promise<void> {
   await fetchQuote({
     fromWallet: ethWallet,
     fromCurrencyCode: 'UNI',
+    fromTokenId: '1f9840a85d5af5bf1d1762f925bdaddc4201f984',
     toWallet: avaxWallet,
     toCurrencyCode: 'JOE',
+    toTokenId: '6e84a6216ea6dacc71ee8e6b0a5b7322eebc0fdd',
     exchangeAmount: '100',
     quoteFor: 'from'
   })
@@ -248,8 +240,10 @@ async function main(): Promise<void> {
   await fetchQuote({
     fromWallet: arrrWallet,
     fromCurrencyCode: 'ARRR',
+    fromTokenId: null,
     toWallet: btcWallet,
     toCurrencyCode: 'BTC',
+    toTokenId: null,
     exchangeAmount: '109',
     quoteFor: 'from'
   })
@@ -257,8 +251,10 @@ async function main(): Promise<void> {
   await fetchQuote({
     fromWallet: btcWallet,
     fromCurrencyCode: 'BTC',
+    fromTokenId: null,
     toWallet: arrrWallet,
     toCurrencyCode: 'ARRR',
+    toTokenId: null,
     exchangeAmount: '0.01',
     quoteFor: 'from'
   })
@@ -267,8 +263,10 @@ async function main(): Promise<void> {
   await fetchQuote({
     fromWallet: ethWallet,
     fromCurrencyCode: 'ETH',
+    fromTokenId: null,
     toWallet: btcWallet,
     toCurrencyCode: 'BTC',
+    toTokenId: null,
     exchangeAmount: '0.004',
     quoteFor: 'to'
   })
@@ -276,8 +274,10 @@ async function main(): Promise<void> {
   await fetchQuote({
     fromWallet: maticWallet,
     fromCurrencyCode: 'USDC',
+    fromTokenId: '3c499c542cef5e3811e1192ce70d8cc03d5c3359',
     toWallet: ethWallet,
     toCurrencyCode: 'WBTC',
+    toTokenId: '2260fac5e5542a773aa44fbcfedf7c193bc2c599',
     exchangeAmount: '4000',
     quoteFor: 'from'
   })
@@ -285,8 +285,10 @@ async function main(): Promise<void> {
   const quote = await fetchQuote({
     fromWallet: bchWallet,
     fromCurrencyCode: 'BCH',
+    fromTokenId: null,
     toWallet: ethWallet,
     toCurrencyCode: 'WBTC',
+    toTokenId: '2260fac5e5542a773aa44fbcfedf7c193bc2c599',
     exchangeAmount: '10',
     quoteFor: 'from'
   })
@@ -296,8 +298,10 @@ async function main(): Promise<void> {
   await fetchQuote({
     fromWallet: bchWallet,
     fromCurrencyCode: 'BCH',
+    fromTokenId: null,
     toWallet: ethWallet,
     toCurrencyCode: 'WBTC',
+    toTokenId: '2260fac5e5542a773aa44fbcfedf7c193bc2c599',
     nativeAmount: quote?.toNativeAmount ?? '0',
     // nativeAmount: await ethWallet.denominationToNative('0.15185834', 'WBTC'),
     quoteFor: 'to'
@@ -306,8 +310,10 @@ async function main(): Promise<void> {
   await fetchQuote({
     fromWallet: btcWallet,
     fromCurrencyCode: 'BTC',
+    fromTokenId: null,
     toWallet: bchWallet,
     toCurrencyCode: 'BCH',
+    toTokenId: null,
     exchangeAmount: '0.2',
     quoteFor: 'from'
   })
@@ -315,8 +321,10 @@ async function main(): Promise<void> {
   await fetchQuote({
     fromWallet: bchWallet,
     fromCurrencyCode: 'BCH',
+    fromTokenId: null,
     toWallet: btcWallet,
     toCurrencyCode: 'BTC',
+    toTokenId: null,
     exchangeAmount: '0.2',
     quoteFor: 'from'
   })